ref: bridge header lookup to ObjC via wrapper class
Expose the case-insensitive header lookup to Objective-C through a thin
@objc SentryHTTPHeaderReader wrapper class instead of an @objc extension
method. Move the extension's tests into HTTPURLResponse+SentryTests.swift.
