+++ Sources/Swift/Networking/HTTPURLResponse+Sentry.swift
@@ -16,16 +16,6 @@ extension HTTPURLResponse {
         return sentryCaseInsensitiveHeaderValue(forName: name)
     }
 
-    /// Objective-C-callable wrapper around `value(forHTTPHeaderFieldCaseInsensitive:)`, so
-    /// Objective-C call sites (e.g. `SentryNetworkTracker`) route header lookups through the same
-    /// sanctioned, case-insensitive helper, including the pre-macOS-10.15 fallback. We can't call
-    /// `-[NSHTTPURLResponse valueForHTTPHeaderField:]` directly because it isn't available before
-    /// macOS 10.15.
-    @objc(sentryValueForHTTPHeaderFieldCaseInsensitive:)
-    @_spi(Private) public func sentryValue(forHTTPHeaderFieldCaseInsensitive name: String) -> String? {
-        value(forHTTPHeaderFieldCaseInsensitive: name)
-    }
-
     /// Pre-macOS-10.15 fallback is an extra method so it can be unit tested. We plan on bumping to macOS 12
     /// in August 2026, so we don't use this implementation for all platforms.
     ///
@@ -34,11 +24,35 @@ extension HTTPURLResponse {
     /// response headers.
     func sentryCaseInsensitiveHeaderValue(forName name: String) -> String? {
         let lowercasedName = name.lowercased()
+        // The linter forbids `allHeaderFields` because its subscript is case-sensitive, and it points
+        // callers to `value(forHTTPHeaderField:)`. That API is only available on macOS 10.15+, so this
+        // pre-10.15 fallback has to iterate `allHeaderFields` instead. The case-sensitivity bug the
+        // linter guards against doesn't apply here: we don't subscript, we compare every key against
+        // `name` with both sides lowercased.
+        // swiftlint:disable avoid_all_header_fields
         for (key, value) in allHeaderFields {
             if let key = key as? String, key.lowercased() == lowercasedName {
                 return value as? String
             }
         }
+        // swiftlint:enable avoid_all_header_fields
         return nil
     }
 }
+
+// swiftlint:disable missing_docs
+
+/// Objective-C bridge for the case-insensitive header lookup. Objective-C can't call the Swift
+/// `HTTPURLResponse` extension method above, so this thin static wrapper exposes it. We can't call
+/// `-[NSHTTPURLResponse valueForHTTPHeaderField:]` from Objective-C directly because it isn't
+/// available before macOS 10.15.
+@objc(SentryHTTPHeaderReader) @_spi(Private)
+public final class HTTPHeaderReader: NSObject {
+
+    @objc(valueForHTTPHeaderFieldCaseInsensitive:inResponse:)
+    public static func value(forHTTPHeaderFieldCaseInsensitive name: String, in response: HTTPURLResponse) -> String? {
+        response.value(forHTTPHeaderFieldCaseInsensitive: name)
+    }
+}
+
+// swiftlint:enable missing_docs
